Fix sourceMapURL when building bundles for windows/macOS (#763)
Summary:
**Summary**

The sourceMapUrl in bundles built for windows/macOS is of the format

```
//# sourceMappingURL=//localhost:8081/index.map?platform=windows&dev=true&hot=false&inlineSourceMap=false
```

This causes the chrome debugger to be unable to correctly load the source map.  It turns out there is already code to modify this specifically for android/iOS.  This change makes windows/macOS behavior match android/iOS.

See https://github.com/microsoft/react-native-windows/issues/9407.

Pull Request resolved: https://github.com/facebook/metro/pull/763

Reviewed By: robhogan

Differential Revision: D51031030

Pulled By: motiz88

fbshipit-source-id: a325e417d725c53c5d96866d80f300a72389e4c2