Respond with HTTP 404 if bundle entry point cannot be resolved
Summary:
Changelog: **[Fix]** Respond with HTTP 404 when a bundle entry point cannot be resolved.

Currently, Metro uses HTTP code 500 (Internal Server Error) for all errors. In the case of "bundle not found" errors, it can be useful to respond with 404 (Not Found) instead. This seemingly was Metro's behaviour at one point (D9772264) but has since regressed.

Reviewed By: huntie

Differential Revision: D51856058

fbshipit-source-id: 63ea7a7cbab5b0347def962867dcc720ac84946b