Refactor Graph - Simplify delta derivation, remove redundant diffing pass
Summary:
D51665495 allows for some simplification of the way `Graph` currently builds its delta for the caller. 

Now that previous `moduleData` is captured on `_releaseModule`, `_processModule` always has a baseline for comparison with the pre-commit graph, and unmodified modules (all dependencies equal, transform result key equal) will now definitely not be marked `touched` even in cases where they are removed and re-added to the graph.

This allows us to cut out much of the additional pruning logic added in D45691844, and reduce some repetition in dependency comparisons.

Changelog: Internal

Reviewed By: motiz88

Differential Revision: D51941768