feat: pass dev option to resolver (#1134)
Summary:
Have first-class support for resolving based on if the process is bundling for `dev` or not. Ref https://discord.com/channels/514829729862516747/514832110595604510/1171554890456256522

This feature is essentially a built-in version of just passing `dev` as a custom resolver option.

Pull Request resolved: https://github.com/facebook/metro/pull/1134

Test Plan: Updated the tests, defaulting to `dev: true`.

Reviewed By: huntie

Differential Revision: D51111860

Pulled By: robhogan

fbshipit-source-id: 40e7482717573489d590ad9b52bf92b9466e08ad