Refactor Graph - mutate #resolvedContexts on edge addition, not resolution
Summary:
`Graph` keeps a private `#resolvedContexts` map with `RequireContext` objects - this is separate from the main `dependencies` `Module` map because it's efficient to be able to iterate over all resolved contexts looking for matches on file addition/deletion.

Currently we update `#resolvedContexts` during dependency resolution. Equivalently, we can collect them during resolution, and update `#resolvedContexts` within `_addDependency`.

This motivation here is that this makes `resolveDependencies` side-effect free, which is necessary for separating traversal from graph mutation, as follows in this stack.

Changelog: Internal

Reviewed By: motiz88

Differential Revision: D51589388

fbshipit-source-id: 274b3e243408dd94d990b38e37c1aa46fc3669de