Remove Fast Refresh bail-out in DeltaCalculator on certain errors
Summary:
Since D51665495, a transform or resolution error encountered while processing modified files will throw *before any change is committed to `Graph`'s state*, so if `DeltaCalculator` catches an error we can be confident `Graph` is still in a good state.

Previously this might've left a modification partially applied to the graph (I assume that's what's meant by "weird state" in the comment here), so `DeltaCalculator` would attempt to detect that and force a reset update. That creates extra work both for Metro (especially as the next modification might not fix the error, we would re-traverse the whole graph on any change perhaps multiple times), and the client, having a full reset to process. It's no longer necessary.

(Note that in any case this safety net had gaps - it was always possible for a bad update to both add and remove modules such that the overall size didn't change.)

Changelog:
```
**[Fix]:** Don't unnecessarily reset Fast Refresh after recovering from certain transform or resolution errors.
```

Reviewed By: GijsWeterings

Differential Revision: D52193122

fbshipit-source-id: 65516257edb0a2cad79c3346f946ab960b7dfc70