FileSystem: Replace getRealPath with lookup - allow for extended information (#1215)
Summary:
Pull Request resolved: https://github.com/facebook/metro/pull/1215

Implement `FileSystem.lookup()`, replacing `FileSystem.getRealPath()`, both internal APIs.

This is motivated by incremental resolution - in particular, we'll need to return more information from `TreeFS` in the context of resolution for both existence (which symlinks have been traversed that would invalidate a resolution if deleted) and non-existence (the first non-existent real path, and the symlinks traversed to get there), so we switch to an extensible object for both cases vs the current `?string`.

Changelog: Internal

Reviewed By: huntie

Differential Revision: D52390401

fbshipit-source-id: 9f2dd765a0f4cc4af25a5002afe231c09cb0614c