TreeFS: Garbage-collect empty directory nodes
Summary:
TreeFS maintains an in-memory representation of the tree of real paths in Metro, using directory nodes (`Map`) that may have directory nodes or file metadata nodes (tuples) as children. 

A "directory" node represents a non-empty directory after Metro's exclusions. The non-existence of a directory node does not imply non-existence of a directory on disk (similar to git or hg), but they are equivalent for the purposes of resolution.

Currently, removing a file deletes the tuple from its parent map, but a map may (~harmlessly) exist with no more children.

This recursively deletes ancestry on deletion of the last file in a directory, so that we only have non-empty nodes (with the possible exception of the root, which is always present). 

This is important for consistently returning the first non-existent node encountered on a lookup, which will be used in the implementation of incremental resolution.

Changelog: Internal

Reviewed By: huntie

Differential Revision: D52391113