Prepare HttpStore for separate configurability for reads and writes
Summary:
Up until now, `HttpStore` has consumed a config where the writes and reads to and from a remote storage API have been using the same setup. We internally need to migrate from the buck1 remote cache to another provider, and Manifold is the easiest option since it provides an HTTP API.

However, unlike the special buckcache endpoint, reads and writes to this API need somewhat different setup. I'm introducing the Manifold configuration later in this stack, but this diff is simply refactoring HttpStore so that it optionally allows for separate configs for reads and writes. Aside from that, it also adds the ability to append headers and URL query parameters. These are both needed for Manifold as well.

All changes have been made in a backwards compatible setup, where we still accept a singular configuration for more homogeneous APIs out there.

Reviewed By: robhogan

Differential Revision: D52917202

fbshipit-source-id: 376a1604c1a7cff75ba297430fee231f9d569f2a