Allow configuring additional success statuses in HttpStore
Summary:
Changelog:

* **[Feature]** Add `additionalSuccessStatuses` option to `HttpStore`.

Adds a mechanism for specifying HTTP response codes to treat as successful (on top of actual success codes like 200) when reading from/writing to a remote cache server.

It can be useful to tell Metro to ignore certain errors coming from an HTTP cache backend, particularly on write: if the error is, for example, that a concurrent write has happened to the same key, it's conceptually safe to ignore - because we can reasonably assume that the write is coming from another instance of Metro that has the same cache configuration.

Reviewed By: GijsWeterings

Differential Revision: D55752523

fbshipit-source-id: f4953d3151e6d74714690bda9cc43d004db91b79