Delete unnecessary Promise libdef override
Summary:
The builtin libdef for Promise now can handle null inputs to catch and then, so the reason to override no longer applies.

My main motivation is to get rid of the `$await` hack, but since we can just delete this file completely, there is no need to replace `$await`.

Changelog: [internal]

Reviewed By: pieterv

Differential Revision: D56860408

fbshipit-source-id: a6e7830e9b91da1198ba47448a7ecdcf0118d29d