Fix --initial-sync by comparing with timeWhenInitiated
Summary:
fix `--initial-sync` not working w/ the previous diff: D58213626

--

in this diff we leveraged querying `mtime_ms` (modified time in milliseconds) when issuing the watchman `subscribe` command, so that in each subscription event we get the information of when the file was modified, and compare that with the epoch time when the watchman subscription was initiated.

if it's prior to the clock, we ignore it.

https://facebook.github.io/watchman/docs/cmd/query#available-fields

--

## a potentially cleaner solution (that didn't work)

I think the current solution is already straight-forward, but I've also tried what I thought could be a cleaner approach by passing the unix epoch time directly as `since`, as the `clockSpec` doc said that you could pass:

> number of seconds since the unix epoch (unix time_t style)

but when passing it directly to:

```
{
  since: {
    clock: String(Math.floor(Date.now().getTime() / 1000)),
    scm: {
      'mergebase-with': 'master',
    },
  },
}
```

then for some reason we cannot initiate the subscription, with incomprehensible message. I'm guessing it's

```
Failed to subscribe:  Error: The watchman connection was closed
    at Client.cancelCommands (/data/sandcastle/boxes/fbsource/xplat/js/node_modules/fb-watchman/index.js:52:15)
    at Socket.<anonymous> (/data/sandcastle/boxes/fbsource/xplat/js/node_modules/fb-watchman/index.js:128:12)
    at Socket.emit (node:events:530:35)
    at endReadableNT (node:internal/streams/readable:1696:12)
    at process.processTicksAndRejections (node:internal/process/task_queues:82:21)
Failed to subscribe:  Error: The watchman connection was closed
    at Client.cancelCommands (/data/sandcastle/boxes/fbsource/xplat/js/node_modules/fb-watchman/index.js:52:15)
    at Socket.<anonymous> (/data/sandcastle/boxes/fbsource/xplat/js/node_modules/fb-watchman/index.js:128:12)
    at Socket.emit (node:events:530:35)
    at endReadableNT (node:internal/streams/readable:1696:12)
    at process.processTicksAndRejections (node:internal/process/task_queues:82:21)
```

Reviewed By: amyogit

Differential Revision: D58339944

fbshipit-source-id: fc98dda7d5f16f5c7fe731e65b57865accb43bf6