Fix non awaited promise in unit test
Summary:
Flow type declarations for expect function has a problem with async arguments.

Whenever we are expecting something on a promise, the expect statement should be awaited, otherwise the result would be incorrect.

For example, this test passes

```
  jest.useRealTimers();
  it('.resolves is ignored when promise is resolved in future', () => {
    const promise = new Promise<number>(resolve =>
      setTimeout(() => resolve(11), 1000),
    );

    expect(promise).resolves.toBe(12);
  });
```

To remove this problem I am going to modify flow type definitions.

D58080193 is a demonstration of this problem

 ---

This diff adds await to async expect statements

Differential Revision: D58398671

fbshipit-source-id: ff7987216233e76d3cc1a5ec56147b7293f8c0f0