Implement require.unstable_importMaybeSync and use it in JSResource (#1296)
Summary:
Pull Request resolved: https://github.com/facebook/metro/pull/1296

This implements `require.unstable_importMaybeSync`, a new API to avoid having to wait for the next tick when importing a value using dynamic `import()` in bundles that don't use bundle splitting.

This API behaves this way:
* If the app is using bundle splitting, `require.unstable_importMaybeSync()` behaves exactly like `import()`. It always returns a promise and it marks a split point for the bundle.
* If the app isn't using bundle splitting, it behaves like a static import (e.g.: `import * as module from 'module'`) and returns the exports for that module synchronously.

This isn't meant to be used by users directly but to build other more specific APIs on top of this.

It includes a small refactor to `asyncRequire.js`.

Reviewed By: motiz88

Differential Revision: D58873729

fbshipit-source-id: 42edd0d61fe36019817a39b136ab0636ad88dd1c