Add run-from-source entry point for metro-config
Summary:
Add an optimistic `unstable_registerForMonorepo` call to the entry point of `metro-config`, allowing it to be the first Metro package loaded when running Metro from source.

In particular, `react-native-community/cli` accesses `metro-config` while loading `react-native.config.js`, so it's often the first Metro package imported when using the community CLI and Metro-from-source.

Reviewed By: hoxyq

Differential Revision: D59376985

fbshipit-source-id: a314f872f42631bbd75841dbfe8c03dae00dd5ea