metro-cache: Implement configurable retry for HttpStore
Summary:
Allow `HttpStore` to be configured with per-endpoint `maxAttempts`, `retryStatuses` and `retryNetworkErrors`, implemented with a jittery exponential backoff.

For most use cases it's likely preferable to ignore transient read errors since a delay will often be slower than a fresh transformation, and similarly failed writes are not generally problematic, but for use cases like warming a remote cache, we'd prefer to fill the cache completely and write no more than necessary.

Reviewed By: GijsWeterings

Differential Revision: D59871321

fbshipit-source-id: 4197d6046eb3c3b4f9ed7f1c1f0bcab4478c08b9