Replace `rimraf` package with `fs.rmSync` (#1311)
Summary:
This is for the deprecation message:
> Rimraf versions prior to v4 are no longer supported deprecation warning.

The `rimraf` package seems to be largely unused in metro, except for `FileStore#_removeDirs`, where it can be replaced with `fs.rmSync`, [which has been added in Node 14.14/16.14/17.3](https://nodejs.org/api/fs.html#fsrmsyncpath-options). Since the minimum supported Node version (via `engines`) is set to `>=18` this seems fine to use.

A `metro-memory-fs` addition of `rmSync` has been added. A comment has been left to note that it's not the most accurate mock method. I'm not sure what the standard of additons to this package are.

Changelog: [Fix] Remove use of `rimraf` and replace it with `fs.rmSync`

Pull Request resolved: https://github.com/facebook/metro/pull/1311

Test Plan:
- `rmSync` has been added to `metro-memory-fs` and tests have been added for it
- Existing tests should cover changes to `FileStore`, since functionality is unchanged

Reviewed By: huntie

Differential Revision: D60778284

Pulled By: robhogan

fbshipit-source-id: c73fcf64103a0efae1020133357db938f259139a