Resolution: Replace context.unstable_getRealPath with context.unstable_fileSystemLookup
Summary:
Replace `unstable_getRealPath`, which was introduced to bring symlink support to the resolver, with the more powerful `unstable_fileSystemLookup`.

We're going to need API(s) that can
 - Check existence of a file or directory
 - Get the real path, to allow resolvers to return real paths.

This API aims to roll everything into one to reduce multiple lookups as much as possible. (I plan to deprecate `context.fileExists` for that reason, because currently using it in practice necessitates a second lookup to determine the real path for a correct resolution.)

This mirrors some previous underlying implementation changes in D52255863 and D52390401 and brings the context API more in line, without exposing more than necessary. As with D52255863, we use `{exists: false}` rather than a `null` return to allow for extended information in a non-breaking change in future - in particular whether the lookup goes outside a watched folder.

Changelog: Internal

Reviewed By: huntie

Differential Revision: D59003947

fbshipit-source-id: 02db458be7701fb8c5fd5feac8eff109ebc6bf1a