Resolver perf: Prefer Maps over dynamic objects for normalised `exports` field (5/n) (#1335)
Summary:
Pull Request resolved: https://github.com/facebook/metro/pull/1335

In V8, dynamic objects (that cannot use a hidden class optimisation) are slow - `Map` is both more performant and generally safer.

By normalising the `exports` field to a `Map` object (once, as normalisations are cached), subsequent lookup, iteration and reduction is considerably faster, improving `resolvePackageTargetFromExports` around 2x.

This is an implementation detail of `PackageExportsResolve.js`, the normalised format is not exposed outside that module.

This brings total resolution time for RNTester down below 100ms, vs >1500ms at the beginning of this stack. (I'll allow myself a 🚀 for that).

Changelog:
```
 - **[Performance]**: Prefer `Map` to dynamic JS objects in package exports resolver.
```

Reviewed By: huntie

Differential Revision: D61850544