Inline requires: cleanup bailout test case
Summary:
Cleanup one of a set of tests intended to check that the input is *not* modified in this particular case.

The standard way to do this is to use `snapshot: false`, which checks that input == output. This test used `snapshot: true` only because the input was not correctly formatted, so the output had whitespace differences.

Allows for clearer grouping of bail-out tests as we introduce memoization later in the stack.

Changelog: Internal

Reviewed By: vzaidman

Differential Revision: D62084971

fbshipit-source-id: 2c5d856cc3a7c64feb919c46093ceb2af8c7d42b