fetch origin main in publish gh action (#1344)
Summary:
A publish just failed: https://github.com/facebook/metro/actions/runs/10699968796 because a tag pushed on the main branch was not validated as being on the main branch.
We assume it's because `main` was not fetched.
Adding
`git fetch origin main`
Which is equivalent to what we do with the hotfix branch a few lines below:
`git fetch origin ${RELEASE_BRANCH}`

Pull Request resolved: https://github.com/facebook/metro/pull/1344

Reviewed By: robhogan

Differential Revision: D62179145

Pulled By: vzaidman

fbshipit-source-id: 9c40d8683d04fdf193c04a9b73daa0e02d190454