Reserve import specifiers beginning '#' for future package imports support
Summary:
We don't have [package.json subpath imports](https://nodejs.org/api/packages.html#subpath-imports) support yet, but we'd like to give ourselves room to implement it as a non-breaking change at some point in Metro 0.81 / React Native 0.76.

This reserves a space at the right place in the resolution algorithm to add an implementation later, according to https://nodejs.org/api/esm.html#resolution-algorithm-specification.

The breaking change here is that `#foo` is no longer a valid package or Haste name, or browser-field key. It will always be interpreted at a subpath import. This is consistent with the rest of the ecosystem and so shouldn't cause any real-world friction.

Changelog:
```
 - **[Breaking]** Resolver: Reserve import specifiers beginning '#' exclusively for future subpath imports support.
```

Reviewed By: huntie

Differential Revision: D64316184

fbshipit-source-id: d48613e66e904812aef7fef0e2e6fe0e54b5f4ff